+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRNNJNFS48W3VRFC4NKQSR0P"
+        "01KRRB9HCQGCHBS2KJRR0FRSH7"
     )
     r"""The unique identifier of the data source"""