+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRNNJNFRJ5CZZ8NPR9RS1ENQ"
+        "01KRRB9HCPRXQ55ZGH201CR7SF"
     )
     r"""The unique identifier of the data source"""