+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -1808,7 +1808,7 @@ class RunAgentAgentToolInputRunTools(BaseModel):
 
     schema_: Annotated[AgentToolInputRunSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KRNNJN1WKSRX8857WSA7HP68"
+    id: Optional[str] = "01KRRB9GWHNP8BVM0SPSH28NDX"
 
     description: Optional[str] = None