+++ packages/orq-rc/src/orq_ai_sdk/models/updateidentityop.py
@@ -194,7 +194,7 @@ class UpdateIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-05-15T11:15:21.069Z")
+    updated: Optional[datetime] = parse_datetime("2026-05-16T12:13:19.225Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")